Move state up
